Fixed mistake in reading GEO coordinates: TSPLIB documentation mandates distances are truncated to int

+++ src/tsplib.h
@@ -367,7 +367,7 @@ namespace as {
                         const auto qa = std::acos(0.5f * q);
                         const auto dist = tsp_earth_radius * qa + 1.0f;
 
-                        return dist;
+                        return (float)(int) dist;
                     };
                 } else if(f_name == "ATT") {
                     return [] (float x1, float y1, float x2, float y2) -> float {
